refactor: seporate formats
